🧪 Add test for unknown tool execution in registry.logic.test.ts
