+++ patch.diff
@@ -1,0 +1,42 @@
+--- src/tools/registry.logic.test.ts
++++ src/tools/registry.logic.test.ts
+@@ -37,4 +37,34 @@
+       isError: true
+     })
+   })
++
++  it('should handle unknown tool execution gracefully', async () => {
++    // Mock server
++    const server = {
++      setRequestHandler: vi.fn()
++    } as any
++
++    // Mock accounts
++    const accounts = [] as any
++
++    // Call registerTools
++    registerTools(server, accounts)
++
++    // Find the handler for CallToolRequestSchema
++    const callToolHandler = server.setRequestHandler.mock.calls.find(
++      (call: any) => call[0] === CallToolRequestSchema
++    )?.[1]
++
++    expect(callToolHandler).toBeDefined()
++
++    // Simulate request with unknown tool name
++    const request = {
++      params: {
++        name: 'unknown_tool_name',
++        arguments: {}
++      }
++    }
++
++    const result = await callToolHandler(request)
++
++    expect(result.isError).toBe(true)
++    expect(result.content).toHaveLength(1)
++    expect(result.content[0].text).toContain('Unknown tool: unknown_tool_name')
++    expect(result.content[0].text).toContain('Available tools: messages, folders, attachments, send, help')
++  })
+ })
